services/products_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
psycopg[binary,pool]==3.2.3
python-dotenv==1.0.1
